Development

This section is provided for devloper who are intrested in extending the functionality of the GeoEfficiency package or just make use of some of its functionality. this softawre is licenced under the MIT licence.

MIT "Expat" License

 Copyright (c) 2019: Mohamed Krar.
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
 in the Software without restriction, including without limitation the rights
 to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 copies of the Software, and to permit persons to whom the Software is
 furnished to do so, subject to the following conditions:
 
 The above copyright notice and this permission notice shall be included in all
 copies or substantial portions of the Software with an appropriate reference to 
 the original work.
 
 THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.

Configuration

found in the source file Config.jl

parameter	describtion	default value

dataFolder	name of the root directory	"GeoEfficiency"

dataDir	root directory	joinpath(homedir(), dataFolder)

integrate	use the package QuadGK to perform integration	begin using QuadGK; QuadGK.quadgk; end

relativeError		1.0E-4

absoluteError		eps(1.0)

resultsFolder	name of the result directory inside the root directory	"results"

max_display	define the default for maximum number of entries shown in the console in batch mode	20 see max_batch

Error System

GeoEfficiency.GeoException — Type.

custom abstract exception that is the parent of all exception in the GeoEfficiency package

source

GeoEfficiency.InValidDetectorDim — Type.

custom exception indicating invalid radiation detector dimensions

source

GeoEfficiency.@validateDetector — Macro.

@validateDetector cond [text]

throw an InValidDetectorDim if cond is false. Message text is optionally displayed upon validation failure.

Examples

julia> @validateDetector iseven(3) "3 is an odd number!"
ERROR: InValidDetectorDim: 3 is an odd number!

julia> @validateDetector isodd(3) "What even are numbers?"

source

GeoEfficiency.NotImplementedError — Type.

custom exception indicating a source to detector geometry which may be valid but not implemented yet

source

GeoEfficiency.@notImplementedError — Macro.

custom macro to throw NotImplementedError exception

source

Console Input

Julia language is quite reach langauge but it seems a good idea thought to collect repeated tasks involing input from console in compact and coustumized to the need function. this section provid two essential functions to deal with inputs from the console. the first:

GeoEfficiency.input — Function.

UnExported

input(prompt::AbstractString = "?: ", incolor::Symbol = :green)

return a string represent the user respond delimited by new line excluding the new line. prompt the user with the massage prompt defaults to ?. wait until the user type its respond and press return. incolor specify the prompt text color, default to $:green$ may take any of the values :black, :blue, :cyan, :green, :light_black, :light_blue, :light_cyan, :light_green, :light_magenta, :light_red, :light_yellow, :magenta, :red, :white, or :yellow.

Color

The effect of color is not allways respected in all teriminals as some color may be simplly 
	ignored by some teriminals.

source

while the secand is a more complex function:

GeoEfficiency.getfloat — Function.

UnExported

getfloat(prompt::AbstractString = "?: ", from::Real = -Inf, to::Real = Inf; KW...)::Float64

prompts the user with the massage prompt defaults to ?: to input a numerical expression evaluate to a numerical value. check that the numerical value is in interval [from, to[ by default [-∞, ∞[ before returning it as a Float64. throws ArgumentError when the given interval is not valid. if the numerical expression fail to evaluated to numerical value or the numerical value is not in the valid interval the function will warn the user and reprompt him to give a valid expresion.

KW arguments

value::AbstractString="nothing" : if provided the function will not ask for input from the console and take it as if it where inputted from the console [for test propose mainly].

lower::Bool=true : whether or not to inculde from as accepted value.

upper::Bool=false : whether or not to inculde to as accepted value.

Note

A blank input (i.e just a return) is considered as being `0.0`. 
	Input from the `console` can be numerical expression not just a number. 
	expression like ``5/2`` ; ``5//2`` ; ``pi`` ; ``π/2`` ; ``exp(2)`` ; ``1E-2 `` ; ``5.2/3`` ; 
	``sin(1)`` ;  ``sin(1)^2`` are all valid expressions.

Examples

julia> getfloat("input a number:", value="3")
3.0

julia> getfloat("input a number:", value="")
0.0

julia> getfloat("input a number:", value="5/2")
2.5

julia> getfloat("input a number:", value="5//2")
2.5

julia> getfloat("input a number:", value="pi")
3.141592653589793

julia> getfloat("input a number:", value="-2")
-2.0

julia> getfloat("input a number:", value="sin(1)^2")
0.7080734182735712

julia> getfloat("input a number:", 1, 5, value="5", upper=true)
5.0

source

Those function are not exported that is normally the user will not need to use them but they are documented here to allow a develober ranked user to make use of them.

Physics Model

Two abstract detector types defined in the package to classify the detectors, the top most super type,

GeoEfficiency.RadiationDetector — Type.

abstract super-supertype of all detectors types

source

any future detector defination should inhert from RadiationDetector. The secand abstract detector Detector is also a sub-type of RadiationDetector but it only acomodates cylinderical type only.

GeoEfficiency.Detector — Type.

Detector

abstract supertype of all detectors types of cylidericalish shapes. also can be used to construct any leaf type.

source

can be used to constract leaf detector.

GeoEfficiency.Detector — Method.

Detector()

construct and return an object of the Detector leaf types (CylDetector, BoreDetector or WellDetector) according to the input from the console.

Note

all required information is acquired from the console and would warn user on invalid data.

source

also it can be used to construct a concrete detector depend on the provided arguments.

GeoEfficiency.Detector — Method.

Detector(CryRadius::Real, CryLength::Real, HoleRadius::Real, HoleDepth::Real)

construct and return well-type, bore-hole or cylindrical detector according to the arguments. it inspect the arguments and call the appropriate leaf type constructor.

Note

if the value(s) of the last argument(s) is\are zero, it acts as a missing argument(s).

see also: CylDetector, BoreDetector, WellDetector.

source

GeoEfficiency.Detector — Method.

Detector(CryRadius::Real)

same as CylDetector(CryRadius::Real).

source

GeoEfficiency.Detector — Method.

Detector(CryRadius::Real, CryLength::Real)

same as CylDetector(CryRadius::Real, CryLength::Real).

source

GeoEfficiency.Detector — Method.

Detector(CryRadius::Real, CryLength::Real, HoleRadius::Real)

same as BoreDetector(CryRadius::Real, CryLength::Real, HoleRadius::Real) except when HoleRadius = 0.0 it acts as CylDetector(CryRadius::Real, CryLength::Real).

source