Batch Mode Input

GeoEfficiency.typeofSrc — Function

typeofSrc()::SrcType

return the current value of the global GeoEfficiency.srcType.

typeofSrc(x::Int)::SrcType

set and return the value of the global GeoEfficiency.srcType corresponding to x.

srcUnknown = -1 also any negative integer treated as so,

srcPoint = 0,

srcLine = 1,

srcDisk = 2,

srcVolume = 3,

srcNotPoint = 4 also any greater than 4 integer treated as so.

GeoEfficiency.setSrcToPoint — Function

setSrcToPoint()::Bool

return whether the source type is a point or not.

setSrcToPoint(yes::Bool)::Bool

return whether the source type is a point or not after setting srcType to srcPoint if yes = true else if yes = false setting it to srcNotPoint if it was not already set to other non-point type (srcDisk, srcLine, srcVolume).

Note

The user can use this function to change the source type any time.

The source type is set the fist time asked for source.

see also: typeofSrc(::Int).

setSrcToPoint(prompt::AbstractString)::Bool

return whether the source type is a point or not. only prompt the user to set the source type if it were not already set before.

see also: typeofSrc(::Int), setSrcToPoint(::Bool).

Warnning

Currently, the source type has no effect but to decide if the source is a point source or a higher dimension source.